orginized the code to seprated mods and deleted the lib file
