+++ src/main.rs
@@ -1,5 +1,10 @@
+mod config;
+mod run;
+mod types;
+mod utils;
+
 fn main() {
-    let config = rsblocks::load_config().unwrap();
-    let blocks = rsblocks::Blocks::new();
-    rsblocks::run(config, blocks);
+    let config = config::load_config().unwrap();
+    let blocks = types::Blocks::new();
+    run::run(config, blocks);
 }
